requirements.txt
# Core AI/ML Libraries
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
sentence-transformers>=2.2.0
torch>=1.9.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
transformers>=4.20.0

# Vector Stores
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0


# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and Logging
structlog>=23.0.0
prometheus-client>=0.19.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
